updated synthetic data generator to add more realism

+++ elecmart/models/GOLD/CORE/DIMENSIONS/gold_dim_date.sql
@@ -1,3 +1,3 @@
-select *, to_char(date, 'YYYYMM') as month_id,
+select *, to_char(date, 'YYYYMM')::INTEGER as month_id,
     year || '-Q' || quarter as year_quarter
  from {{ref('silver_dim_date')}}
